+++ apps/dx/docs-docusaurus/iac/variables.tf
@@ -1,5 +1,0 @@
-variable "vercel_api_token" {
-  description = "Vercel API token"
-  type        = string
-  sensitive   = true
-}